feat: add timestamps for lip status
We add timestamps for the various lips statuses so that we get better
data on the lip movement. The no-show status gets deprecated as a
displayable value, but will recorded in the timestamp whenever a lip is
moved away from the live stack after less than the NO_SHOW_THRESHOLD
time.